edge_processor/requirements.txt
paho-mqtt
joblib
pandas
scikit-learn
xgboost
